+++ libs/pages/services/src/lib/feature/page-application-create-feature/step-general-feature/step-general-feature.spec.tsx
@@ -1,0 +1,26 @@
+import { render } from '__tests__/utils/setup-jest'
+import { ServiceTypeEnum } from '@qovery/shared/enums'
+import { ApplicationContainerCreateContext } from '../page-application-create-feature'
+import StepGeneralFeature from './step-general-feature'
+
+describe('PageApplicationCreateGeneralFeature', () => {
+  it('should render successfully', () => {
+    const { baseElement } = render(
+      <ApplicationContainerCreateContext.Provider
+        value={{
+          currentStep: 1,
+          setCurrentStep: jest.fn(),
+          generalData: { name: 'test', serviceType: ServiceTypeEnum.APPLICATION },
+          setGeneralData: jest.fn(),
+          resourcesData: undefined,
+          setResourcesData: jest.fn(),
+          portData: undefined,
+          setPortData: jest.fn(),
+        }}
+      >
+        <StepGeneralFeature />
+      </ApplicationContainerCreateContext.Provider>
+    )
+    expect(baseElement).toBeTruthy()
+  })
+})
