--- libs/pages/services/src/lib/feature/page-database-create-feature/page-database-create-resources-feature/page-database-create-resources-feature.tsx
@@ -1,80 +1,0 @@
-import { useEffect } from 'react'
-import { FormProvider, useForm } from 'react-hook-form'
-import { useNavigate, useParams } from 'react-router-dom'
-import {
-  SERVICES_DATABASE_CREATION_GENERAL_URL,
-  SERVICES_DATABASE_CREATION_POST_URL,
-  SERVICES_DATABASE_CREATION_URL,
-  SERVICES_URL,
-} from '@qovery/shared/router'
-import { FunnelFlowBody, FunnelFlowHelpCard } from '@qovery/shared/ui'
-import { useDocumentTitle } from '@qovery/shared/utils'
-import PageDatabaseCreateResources from '../../../ui/page-database-create/page-database-create-resources/page-database-create-resources'
-import { ResourcesData } from '../database-creation-flow.interface'
-import { useDatabaseCreateContext } from '../page-database-create-feature'
-
-export function PageDatabaseCreateResourcesFeature() {
-  useDocumentTitle('Resources - Create Database')
-  const { setCurrentStep, resourcesData, setResourcesData, generalData } = useDatabaseCreateContext()
-  const { organizationId = '', projectId = '', environmentId = '' } = useParams()
-  const navigate = useNavigate()
-  const pathCreate = `${SERVICES_URL(organizationId, projectId, environmentId)}${SERVICES_DATABASE_CREATION_URL}`
-
-  useEffect(() => {
-    !generalData?.name &&
-      navigate(
-        `${SERVICES_URL(organizationId, projectId, environmentId)}${SERVICES_DATABASE_CREATION_URL}` +
-          SERVICES_DATABASE_CREATION_GENERAL_URL
-      )
-  }, [generalData, navigate, environmentId, organizationId, projectId])
-
-  const funnelCardHelp = (
-    <FunnelFlowHelpCard
-      title="Setting the right resources"
-      items={[
-        'Databases are deployed as containers on your Kubernetes cluster',
-        'Set the vCPU/RAM based on your application need',
-        'You can’t go beyond the resources available on your cluster',
-      ]}
-      helpSectionProps={{
-        description: 'Need help? You may find these links useful',
-        links: [
-          {
-            link: 'https://hub.qovery.com/docs/using-qovery/configuration/database/#resources',
-            linkLabel: 'How to configure my database',
-            external: true,
-          },
-          { link: 'https://discuss.qovery.com/', linkLabel: 'Still need help? Ask on our Forum', external: true },
-        ],
-      }}
-    />
-  )
-
-  useEffect(() => {
-    setCurrentStep(2)
-  }, [setCurrentStep])
-
-  const methods = useForm<ResourcesData>({
-    defaultValues: resourcesData,
-    mode: 'onChange',
-  })
-
-  const onSubmit = methods.handleSubmit((data) => {
-    setResourcesData(data)
-    navigate(pathCreate + SERVICES_DATABASE_CREATION_POST_URL)
-  })
-
-  const onBack = () => {
-    navigate(pathCreate + SERVICES_DATABASE_CREATION_GENERAL_URL)
-  }
-
-  return (
-    <FunnelFlowBody helpSection={funnelCardHelp}>
-      <FormProvider {...methods}>
-        <PageDatabaseCreateResources onBack={onBack} onSubmit={onSubmit} />
-      </FormProvider>
-    </FunnelFlowBody>
-  )
-}
-
-export default PageDatabaseCreateResourcesFeature
