+++ libs/pages/services/src/lib/ui/page-application-create/step-general/step-general.tsx
@@ -1,0 +1,131 @@
+import { FormEventHandler } from 'react'
+import { Controller, useFormContext } from 'react-hook-form'
+import { useNavigate, useParams } from 'react-router-dom'
+import {
+  CreateGeneralGitApplication,
+  EntrypointCmdInputs,
+  GeneralContainerSettings,
+} from '@qovery/shared/console-shared'
+import { IconEnum, ServiceTypeEnum, isApplication, isContainer } from '@qovery/shared/enums'
+import { ApplicationGeneralData, OrganizationEntity } from '@qovery/shared/interfaces'
+import { SERVICES_URL } from '@qovery/shared/router'
+import { Button, ButtonSize, ButtonStyle, Icon, InputSelect, InputText, InputTextArea } from '@qovery/shared/ui'
+
+export interface StepGeneralProps {
+  onSubmit: FormEventHandler<HTMLFormElement>
+  organization?: OrganizationEntity
+}
+
+export function StepGeneral(props: StepGeneralProps) {
+  const { control, watch, formState } = useFormContext<ApplicationGeneralData>()
+  const { organizationId = '', environmentId = '', projectId = '' } = useParams()
+  const navigate = useNavigate()
+
+  const watchServiceType = watch('serviceType')
+
+  return (
+    <div>
+      <div className="mb-10">
+        <h3 className="text-text-700 text-lg mb-2">General informations</h3>
+        <p className="text-text-500 text-sm mb-2">
+          General settings allow you to set up your application name, git repository or container settings.
+        </p>
+      </div>
+
+      <form onSubmit={props.onSubmit}>
+        <Controller
+          name="name"
+          control={control}
+          rules={{
+            required: 'Please enter a name.',
+          }}
+          render={({ field, fieldState: { error } }) => (
+            <InputText
+              className="mb-3"
+              name={field.name}
+              onChange={field.onChange}
+              value={field.value}
+              label="Application name"
+              error={error?.message}
+            />
+          )}
+        />
+        <Controller
+          name="description"
+          control={control}
+          render={({ field }) => (
+            <InputTextArea
+              className="mb-3"
+              name={field.name}
+              onChange={field.onChange}
+              value={field.value}
+              label="Description"
+            />
+          )}
+        />
+        <Controller
+          name="serviceType"
+          control={control}
+          rules={{
+            required: 'Please select a source.',
+          }}
+          render={({ field, fieldState: { error } }) => (
+            <InputSelect
+              dataTestId="input-select-source"
+              className="mb-6"
+              onChange={field.onChange}
+              value={field.value}
+              options={[
+                {
+                  value: ServiceTypeEnum.APPLICATION,
+                  label: 'Git provider',
+                  icon: <Icon name={IconEnum.GIT} className="w-4" />,
+                },
+                {
+                  value: ServiceTypeEnum.CONTAINER,
+                  label: 'Container Registry',
+                  icon: <Icon name={IconEnum.CONTAINER} className="w-4" />,
+                },
+              ]}
+              label="Application source"
+              error={error?.message}
+            />
+          )}
+        />
+
+        <div className="border-b border-b-element-light-lighter-400 mb-6"></div>
+        {isApplication(watchServiceType) && <CreateGeneralGitApplication />}
+
+        {isContainer(watchServiceType) && (
+          <>
+            <GeneralContainerSettings organization={props.organization} />
+            <EntrypointCmdInputs />
+          </>
+        )}
+
+        <div className="flex justify-between">
+          <Button
+            onClick={() => navigate(SERVICES_URL(organizationId, projectId, environmentId))}
+            type="button"
+            className="btn--no-min-w"
+            size={ButtonSize.XLARGE}
+            style={ButtonStyle.STROKED}
+          >
+            Cancel
+          </Button>
+          <Button
+            dataTestId="button-submit"
+            type="submit"
+            disabled={!formState.isValid}
+            size={ButtonSize.XLARGE}
+            style={ButtonStyle.BASIC}
+          >
+            Continue
+          </Button>
+        </div>
+      </form>
+    </div>
+  )
+}
+
+export default StepGeneral
