--- libs/pages/services/src/lib/ui/page-database-create/page-database-create-resources/page-database-create-resources.tsx
@@ -1,49 +1,0 @@
-import { FormEventHandler } from 'react'
-import { useFormContext } from 'react-hook-form'
-import { SettingResources } from '@qovery/shared/console-shared'
-import { Button, ButtonSize, ButtonStyle } from '@qovery/shared/ui'
-import { ResourcesData } from '../../../feature/page-database-create-feature/database-creation-flow.interface'
-
-export interface PageDatabaseCreateResourcesProps {
-  onBack: () => void
-  onSubmit: FormEventHandler<HTMLFormElement>
-}
-
-export function PageDatabaseCreateResources(props: PageDatabaseCreateResourcesProps) {
-  const { formState } = useFormContext<ResourcesData>()
-
-  return (
-    <>
-      <div className="mb-10">
-        <h3 className="text-text-700 text-lg mb-2">Set resources</h3>
-      </div>
-
-      <form onSubmit={props.onSubmit}>
-        <SettingResources isDatabase displayWarningCpu={false} />
-
-        <div className="flex justify-between">
-          <Button
-            onClick={props.onBack}
-            className="btn--no-min-w"
-            type="button"
-            size={ButtonSize.XLARGE}
-            style={ButtonStyle.STROKED}
-          >
-            Back
-          </Button>
-          <Button
-            dataTestId="button-submit"
-            type="submit"
-            disabled={!formState.isValid}
-            size={ButtonSize.XLARGE}
-            style={ButtonStyle.BASIC}
-          >
-            Continue
-          </Button>
-        </div>
-      </form>
-    </>
-  )
-}
-
-export default PageDatabaseCreateResources
