+++ libs/shared/console-shared/src/index.ts
@@ -10,8 +10,8 @@ export * from './lib/buttons-actions/ui/application-buttons-actions/application-
 export * from './lib/buttons-actions/ui/database-buttons-actions/database-buttons-actions'
 export * from './lib/buttons-actions/ui/cluster-buttons-actions/cluster-buttons-actions'
 export * from './lib/create-clone-environment-modal/feature/create-clone-environment-modal-feature'
-export * from './lib/flow-create-port/ui/flow-create-port'
-export * from './lib/flow-create-variable/ui/flow-create-variable'
+export * from './lib/flow-create-port/ui/flow-create-port/flow-create-port'
+export * from './lib/flow-create-variable/ui/flow-create-variable/flow-create-variable'
 export * from './lib/create-general-git-application/ui/create-general-git-application'
 export * from './lib/git-repository-settings/feature/edit-git-repository-settings-feature/edit-git-repository-settings-feature'
 export * from './lib/create-project-modal/feature/create-project-modal-feature'
